vk_weather_bot_render/requirements.txt
vk-api
requests
python-dotenv